pulling out more of the code for node loading
