+++ core/src/main/java/org/neo4j/graphalgo/core/heavyweight/Nodes.java
@@ -59,7 +59,7 @@ public class Nodes {
         return offset;
     }
 
-    public long rows() {
+    long rows() {
         return rows;
     }
 }
